+++ src/modules/clickOutside/modules/clickOutside.module.ts
@@ -1,0 +1,24 @@
+import {NgModule} from "@angular/core";
+
+import {ClickOutsideDirective} from '../directives/clickOutside/clickOutside.directive';
+
+/**
+ * Module for ClickOutside directive
+ */
+@NgModule(
+{
+    imports:
+    [
+    ],
+    declarations:
+    [
+        ClickOutsideDirective
+    ],
+    exports:
+    [
+        ClickOutsideDirective
+    ]
+})
+export class ClickOutsideModule
+{
+}